requirements/base.txt
Django==1.6.4
South==0.8.4
django-braces==1.4.0
django-extensions==1.3.5
django-model-utils==2.0.3
django-tastypie==0.11.0
ipython==2.0.0
logutils==0.3.3